+++ Plutarch/Rational.hs
@@ -16,7 +16,7 @@ module Plutarch.Rational (
 import Data.Kind (Type)
 import Data.Ratio (denominator, numerator)
 import GHC.Generics (Generic)
-import Plutarch.Bool (POrd, PPartialOrd, pif, pnot, (#<), (#<=))
+import Plutarch.Bool (pif, pnot)
 import Plutarch.Builtin (
   PAsData,
   PBuiltinList,
@@ -32,6 +32,7 @@ import Plutarch.Builtin (
 import Plutarch.Integer (PInteger, PIntegral (pquot), pdiv, pmod)
 import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Lift (pconstant)
+import Plutarch.Internal.Ord (POrd, PPartialOrd ((#<), (#<=)))
 import Plutarch.Internal.Other (pfix, pto)
 import Plutarch.Internal.PLam (plam)
 import Plutarch.Internal.PlutusType (