+++ Plutarch/Unit.hs
@@ -4,12 +4,6 @@
 
 module Plutarch.Unit (PUnit (..)) where
 
-import Plutarch.Bool (
-  POrd,
-  PPartialOrd,
-  (#<),
-  (#<=),
- )
 import Plutarch.Builtin.Bool (PBool (PFalse, PTrue))
 import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Lift (
@@ -18,6 +12,7 @@ import Plutarch.Internal.Lift (
   PLifted (PLifted),
   pconstant,
  )
+import Plutarch.Internal.Ord (POrd, PPartialOrd ((#<), (#<=)))
 import Plutarch.Internal.PlutusType (PInner, PlutusType, pcon, pcon', pmatch')
 import Plutarch.Internal.Term (Term, plet)
 import Plutarch.Show (PShow (pshow'))